+++ lib/roo_on_rails/checks/environment.rb
@@ -4,13 +4,19 @@ require 'roo_on_rails/checks/heroku/app_exists'
 require 'roo_on_rails/checks/heroku/preboot_enabled'
 require 'roo_on_rails/checks/heroku/app_exists'
 require 'roo_on_rails/checks/sidekiq/settings'
+require 'roo_on_rails/checks/heroku/drains_metrics'
+require 'roo_on_rails/checks/documentation/playbook'
+require 'roo_on_rails/checks/google_oauth/initializer'
 
 module RooOnRails
   module Checks
     class Environment < EnvSpecific
       requires GitHub::BranchProtection
+      requires Heroku::DrainsMetrics
       requires Heroku::PrebootEnabled
       requires Sidekiq::Settings
+      requires Documentation::Playbook
+      requires GoogleOauth::Initializer
 
       def call
         # nothing to do