+++ lib/roo_on_rails/railties/database.rb
@@ -1,8 +1,8 @@
 module RooOnRails
   module Railties
     class Database < Rails::Railtie
-      if ActiveRecord::VERSION::MAJOR >= 4
-        initializer 'roo_on_rails.database', after: 'active_record.initialize_database' do
+      initializer 'roo_on_rails.database', after: 'active_record.initialize_database' do
+        ActiveSupport.on_load :active_record do
           $stderr.puts 'initializer roo_on_rails.database'
 
           config = ActiveRecord::Base.configurations[Rails.env]